+++ systems/identity/identity-application/src/main/kotlin/hs/kr/entrydsm/identity/application/service/IdentityResultMapper.kt
@@ -4,7 +4,7 @@ import hs.kr.entrydsm.identity.application.port.`in`.result.AccountResult
 import hs.kr.entrydsm.identity.application.port.`in`.result.ApplicationStatusResult
 import hs.kr.entrydsm.identity.application.port.`in`.result.ProfileResult
 import hs.kr.entrydsm.identity.application.port.out.data.ApplicationSnapshot
-import hs.kr.entrydsm.identity.domain.Account
+import hs.kr.entrydsm.identity.domain.model.Account
 
 internal fun Account.toAccountResult(): AccountResult =
     AccountResult(
